Adjusts in calculation

+++ backend/src/main/java/br/com/unicesumar/piscinalimpa/controller/utils/FormulaReplace.java
@@ -2,7 +2,7 @@ package br.com.unicesumar.piscinalimpa.controller.utils;
 
 public class FormulaReplace {
 
-    public static String replaceMultiplierAndVolumeIn(String formula, Integer multiplier, Double volume) {
+    public static String replaceMultiplierAndVolumeIn(String formula, Double multiplier, Double volume) {
         formula = formula.replace("m", multiplier.toString());
         formula = formula.replace("v", volume.toString());
         return formula;
